Red Tie
This repository contains a XHTML file and a version of it adapted for
HTML 5. While I was taking an introductory HTML 5 course one of my
teachers mentioned how much he liked this template. So, I adapted it
using what I learned in HTML 5. No div or span tags are used and the use
of wrapper elements has been all but eliminated. This could be easily
adapted as a HTML PHP or ASP template.